% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport_raws_profile.R
\docType{import}
\name{aws_version}
\alias{aws_version}
\title{AWS CLI version}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{raws.profile}{\code{\link[raws.profile]{aws_version}}}
}}

