% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_topic.R
\name{describe_topic}
\alias{describe_topic}
\title{Topic description}
\usage{
describe_topic(topic, profile = "default")
}
\arguments{
\item{topic}{A topic name or topic ARN.}

\item{profile}{Your profile name in AWS. If profile was not supplied
this function will use the default profile.}
}
\description{
This function returns a tibble with some informations about the topic.
}
\examples{
\dontrun{
describe_topic("raws_sns_test", profile = "your profile")
}

}
