% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_message.R
\name{send_message}
\alias{send_message}
\title{Send message to endpoints}
\usage{
send_message(topic = NULL, phone_number = NULL, message,
  subject = NULL, suppress_output = FALSE, profile = "default")
}
\arguments{
\item{topic}{A topic name or topic ARN.}

\item{phone_number}{Send a message directly to this phone number. This number
doesn't need to be in a topic.}

\item{message}{The massage you want to send.}

\item{subject}{(optional) The message subject. Only available to "email"
protocol.}

\item{suppress_output}{Use TRUE if you don't want the output.}

\item{profile}{Your profile name in AWS. If profile was not supplied
this function will use the default profile.}
}
\description{
You can send messages for every endpoint in a topic or directly to a phone
number. SMS is only available for profiles in us-east-1 region, you can verify
your region profile using \code{raws.profile::profile_settings("your profile")}.
}
\examples{
\dontrun{
# send a message to a topic
send_message(topic = "your_topic",
             subject = "testing",
             message = "This is a test message.")

# send a message directly to a phone number
send_message(phone_number = "+0081345678",
             message = "This is a test message.")
}

}
