% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_.R
\name{sns_lists}
\alias{sns_lists}
\alias{list_topics}
\alias{list_subscriptions}
\title{List your topics and subscriptions}
\usage{
list_topics(profile = "default")

list_subscriptions(topic, profile = "default")
}
\arguments{
\item{profile}{Your profile name in AWS. If profile was not supplied
this function will use the default profile.}

\item{topic}{A topic name or topic ARN.}
}
\description{
The \code{list_topics()} shows the topics created in your SNS
account. You can see your subscriptions using
 \code{list_subscriptions("your topic")}.
}
\examples{
\dontrun{
# list the topics in your account
list_topics()

# list the subscriptions in a topic using a profile
list_subscriptions(topic = "your topic",
                   profile = "your profile")

}
}
