% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{subscription}
\alias{subscription}
\alias{subscribe}
\alias{unsubscribe}
\title{Subscribe and unsubscrive endpoints}
\usage{
subscribe(topic, protocol, endpoint, profile = "default")

unsubscribe(topic, endpoint, subscription_arn, profile = "default")
}
\arguments{
\item{topic}{A topic name or topic ARN.}

\item{protocol}{A string or a vector of strings with the protocol.
"sms" and "email" are supported. If you have a vector of endpoints with the
same protocol you just need to specify the procotol, not a vector of
protocols.}

\item{endpoint}{A string or a vector of strings with users that you want
to subscribe. Use a valid email and for SMS use a complete cell phone number,
for example: "+558112345678".}

\item{profile}{Your profile name in AWS. If profile was not supplied
this function will use the default profile.}

\item{endpoint_arn}{A string with the user's ARN. Use this for better
performance in \code{unsubscribe()}. You can check the endpoint ARN in SNS
account or usind the \code{list_subscription()}}
}
\description{
You can subscribe one or more users with \code{subscribe()}. For unsubscribe
a user use an endpoint_arn for better performance.
}
\examples{
\dontrun{
# subscribe a user
# subscribe(topic = "your topic",
             protocol = "sms",
             endpoint = "+0012345")

# subscribe more than one user with a singe protocol
endpoint <- c("user1@email.com", "user2@email.com", "user3@email.com" )
subscribe(topic = "your topic",
          protocol = "email",
          endpoint = endpoint)

# subscribe more than one user with differents protocols
endpoint <- c("user4@email.com", "+0054321", "user5@email.com" )
protocol <- c("email", "sms", "email")
subscribe(topic = "your topic",
         protocol = protocol,
         endpoint = endpoint)

# unsubscribe a user
unsubscribe(topic = "your topic",
            endpoint = "+0012345")

# unsubscribe more than one user
unsubscribe(topic = "your topic",
            endpoint = c("user4@email.com", "+0054321"))

# use the endpoint ARN for better performance
unsubscribe(topic = "your topic",
            endpoint_arn = "arn:aws:sns:your_topic:12345")
}
}
